==> 4-demo.ipynb <==
from transformers import AutoModelForCausalLM, AutoTokenizer
import torch

model_id = "aisingapore/sea-lion-7b-instruct"
tokenizer = AutoTokenizer.from_pretrained(model_id, trust_remote_code=True)
if tokenizer.pad_token is None:
  tokenizer.pad_token = tokenizer.eos_token

model = AutoModelForCausalLM.from_pretrained(model_id, device_map="auto", trust_remote_code=True, load_in_4bit=True, bnb_4bit_compute_dtype=torch.float16)

Sentimen Positif
- "Hari ini cuaca sangat cerah, dan itu membuat perjalanan ke kantor terasa menyenangkan."
- "Saya baru saja menerima umpan balik yang sangat positif dari klien saya, dan itu memberi saya semangat baru."
- "Makanan di restoran baru yang saya coba benar-benar lezat, dan pelayanan mereka sangat ramah."
- "Saya baru saja menyelesaikan proyek besar di tempat kerja, dan rekan-rekan saya memberi pujian atas kerja keras saya."
- "Liburan akhir pekan ini benar-benar menyegarkan, dan saya merasa siap untuk menghadapi minggu depan."

Sentimen Negatif
- "Saya mengalami kemacetan parah di jalan menuju kantor, dan itu membuat saya terlambat untuk pertemuan penting."
- "Saya mendapatkan ulasan negatif dari atasan saya yang membuat saya merasa sangat kecewa dan kurang percaya diri."
- "Makanan di restoran yang saya kunjungi ternyata tidak sesuai dengan harapan, dan pelayanannya juga sangat lambat."
- "Proyek yang saya kerjakan mengalami banyak kendala, dan rekan-rekan saya mengeluhkan hasilnya."
- "Cuaca panas terus-menerus membuat akhir pekan ini menjadi sangat malas melakukan apapun dan tidak menyenangkan."

text = "Cuaca panas terus-menerus membuat akhir pekan ini menjadi sangat malas melakukan apapun dan tidak menyenangkan."

prompt_template = """### USER: Ubah sentimen dari kalimat awal berikut menjadi sentimen sebaliknya
Kalimat Awal: {text}
Kalimat Baru:

### RESPONSE:
"""

full_prompt = prompt_template.format(text=text)

tokens = tokenizer(full_prompt, return_tensors="pt")
output = model.generate(input_ids=tokens["input_ids"], max_new_tokens=30, eos_token_id=tokenizer.eos_token_id, pad_token_id=tokenizer.pad_token_id, num_return_sequences=1)
response = tokenizer.decode(output[0], skip_special_tokens=True)
print(response)
